requirements.txt
alembic==1.13.2
amqp==5.2.0
aniso8601==9.0.1
async-timeout==4.0.3
backports.zoneinfo==0.2.1
billiard==4.2.0
blinker==1.5
cachelib==0.9.0
celery==5.3.5
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.3
Flask-Admin==1.6.1
Flask-Caching==2.1.0
Flask-Excel==0.0.7
Flask-Login==0.6.2
flask-mail==0.10.0
Flask-Migrate==4.0.7
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.1.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.0.0
importlib-resources==6.4.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.4.2
kombu==5.3.3
lml==0.1.0
Mako==1.3.5
MarkupSafe==2.1.2
mutagen==1.47.0
nltk==3.8.1
numpy==1.24.4
pandas==2.0.3
passlib==1.7.4
pdfkit==1.0.0
pillow==10.4.0
prompt-toolkit==3.0.40
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
regex==2024.7.24
reportlab==4.2.2
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.6
texttable==1.7.0
tqdm==4.66.5
typing-extensions==4.5.0
tzdata==2023.3
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.9
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0
